refactor(store): bump dgeni and move models back to creator files
