+++ projects/ngrx.io/package.json
@@ -109,7 +109,7 @@
     "cross-spawn": "^5.1.0",
     "css-selector-parser": "^1.3.0",
     "dgeni": "^0.4.14",
-    "dgeni-packages": "^0.29.2",
+    "dgeni-packages": "^0.30.0",
     "entities": "^1.1.1",
     "eslint": "^3.19.0",
     "eslint-plugin-jasmine": "^2.2.0",
